id type changed to Long

+++ src/main/java/com/scribblemate/dto/CollaboratorDto.java
@@ -11,8 +11,8 @@ import lombok.Setter;
 @AllArgsConstructor
 public class CollaboratorDto {
 
-	private Integer id;
-	
+	private Long id;
+
 	private String name;
 
 	private String email;
